move localStorage func to util
